+++ R/main_effects.R
@@ -2,7 +2,7 @@
 #'
 #' @param design Design of experiment (Factorial Design)
 #' @param response A character string indicating the response of the data
-#' @param ncols number of columns for facet grid. Default is 2
+#' @param n_columns number of columns for facet grid. Default is 2
 #' @param exclude_vars A vector containing variables to exclude
 #' @param color_palette A character string indicating the color map option to use. Eight options are available: "viridis","cividis","magma","inferno","plasma","rocket","mako","turbo"
 #' @param alpha The alpha transparency, a number in [0,1]
@@ -18,8 +18,8 @@
 #'
 #' @examples
 #' main_effects(original_epitaxial,response='s2',exclude_vars = c('ybar','lns2'))
-#' main_effects(original_epitaxial,response='ybar',exclude_vars=c('A','s2','lns2'),ncols=3)
-main_effects <- function(design,response,ncols=2,
+#' main_effects(original_epitaxial,response='ybar',exclude_vars=c('A','s2','lns2'),n_columns=3)
+main_effects <- function(design,response,n_columns=2,
                          exclude_vars=c(),
                          color_palette = NA,
                          alpha=1,direction = 1,
@@ -83,7 +83,7 @@ main_effects <- function(design,response,ncols=2,
             panel.grid.major = element_blank(),
             panel.grid.minor = element_blank()) +
       scale_color_manual(values = factor_colors)+
-      facet_wrap(vars(ind),ncol = ncols)+
+      facet_wrap(vars(ind),ncol = n_columns)+
       labs(y=paste0("Mean of ",response),x='')
 
     return(p)